Fix HTMX form handling.
We now handle errors in HTMX form submissions, too. Explicitly mark the model attribute we pull the form from to make sure the Errors instance gets populated for the correct form.

Switch to programatic trigger definitions to make sure we only fire the trigger in the case of form success. Renamed the event to entryAdded.

Added hx-oob-swap to the form declaration to override the swap target in case we return the form in turn to display orders.